+++ runtimes/common/src/constants.rs
@@ -363,14 +363,19 @@ pub mod governance {
 	}
 }
 
+pub mod timestamp {
+	use super::*;
+
+	#[allow(clippy::integer_division)]
+	pub const MINIMUM_PERIOD: u64 = SLOT_DURATION / 2;
+}
+
 pub mod multisig {
 	use super::*;
 
-	parameter_types! {
-		pub const MaxSignitors: u32 = 64;
-		pub const DepositBase: Balance = DEPOSIT_STORAGE_ITEM;
-		pub const DepositFactor: Balance = DEPOSIT_STORAGE_BYTE;
-	}
+	pub const MAX_SIGNITORS: u32 = 64;
+	pub const DEPOSIT_BASE: Balance = DEPOSIT_STORAGE_ITEM;
+	pub const DEPOSIT_FACTOR: Balance = DEPOSIT_STORAGE_BYTE;
 }
 
 pub mod pallet_migration {
